Adding python files
Here are the scripts for publishing JSON data & Subscribing to a particular thing & receive data.

ei_aws_publish.py script is used to publish, mac id, random-number, and a random string for every 5 seconds. The ei_aws-subscribe.py script is used to subscribe to a particular topic and receive the published data on that topic.